Add files via upload
I looked over the specifications for the project. Let me know if you agree with this rough plan. I find it easiest if we draw out what we want to see and then program it. At least for the charts and the creation of a new variable. I have 4 charts, one function and some tables.